pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "hydratext"
authors = [{name = "Diogo Vieira", email = "jdmartinsvieira63@gmail.com"},{name = "Eduardo Magalhães", email = "eduardomagalhaes107@gmail.com1"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]

dependencies = [ 
    "streamlit >= 1.20.0",
    "pytesseract >= 0.3.10",
    "Pillow >= 9.5.0",
    "gramformer >= 1.0",
    "nltk >= 3.8.1",
    "pdftotext >= 2.2.2",
    "wget >= 3.2",
    "beautifulsoup4 >= 4.11.2",
    "urllib3 >= 1.26.5",
]

[project.scripts]
hydratxt = "hydratext:hydratext"


